fixed type.

+++ tensorflow_quantum/core/ops/tfq_simulate_samples_op.cc
@@ -224,7 +224,7 @@ class TfqSimulateSamplesOp : public tensorflow::OpKernel {
       sv.release();
     };
 
-    const int num_cycles = 200 * (1 << max_num_qubits);
+    const int64_t num_cycles = 200 * (1 << max_num_qubits);
     context->device()->tensorflow_cpu_worker_threads()->workers->ParallelFor(
         fused_circuits.size(), num_cycles, DoWork);
   }
